split prepare from fetch, global config
