--- firedata/config.ini
@@ -1,14 +1,0 @@
-[OS]
-data_path=/home/tadas/activefire/firedata/data
-[MODIS]
-base_url=https://nrt4.modaps.eosdis.nasa.gov/api/v2/content/archives/FIRMS/modis-c6.1/Global/MODIS_C6_1_Global_MCD14DL_NRT_
-nrt_dataset=nrt_complete.parquet
-archive_end=2021-12-31T23:59
-[VIIRS_NPP]
-base_url=https://nrt4.modaps.eosdis.nasa.gov/api/v2/content/archives/FIRMS/suomi-npp-viirs-c2/Global/SUOMI_VIIRS_C2_Global_VNP14IMGTDL_NRT_
-nrt_dataset=nrt_complete_viirs_npp.parquet
-archive_end=2021-12-31T23:59
-[VIIRS_NOAA]
-base_url=https://nrt4.modaps.eosdis.nasa.gov/api/v2/content/archives/FIRMS/noaa-20-viirs-c2/Global/J1_VIIRS_C2_Global_VJ114IMGTDL_NRT_
-nrt_dataset=nrt_complete_viirs_noaa.parquet
-archive_end=2021-12-31T23:59
